+++ src/template/profilePost/PetPost.jsx
@@ -57,9 +57,6 @@ export function PetPost() {
             )
           })
         }
-        {
-
-        }
       </MiniPostWrap>
     </SectionAllWrap>
   )
